Set up 'CandidRpcClient' for reusability in unit tests
